testutil: Introduce package and move fs fakes there
